Refactor grunt and helpers into separate files as per issue #3, basic modification for grunt 0.4 compatibility
