+++ GruntFile.js
@@ -33,5 +33,5 @@ module.exports = function(grunt) {
   grunt.loadNpmTasks("grunt-contrib-jshint");
   grunt.loadNpmTasks("grunt-contrib-watch");
 
-  grunt.registerTask("default", ["lint", "nodeunit"]);
+  grunt.registerTask("default", ["jshint", "nodeunit"]);
 };
